refactor(transport): move AggregatedRemoteStream to own class and make it a record
